Move ExperimentEncoder and Decoder to framework
