+++ test/randomized_benchmarking/test_rb_analysis.py
@@ -24,8 +24,7 @@ from qiskit.circuit.library import (
 )
 from qiskit_experiments.framework import ExperimentData
 from qiskit_experiments.library import StandardRB, InterleavedRB
-from qiskit_experiments.database_service.json import ExperimentDecoder
-from qiskit_experiments.database_service.db_fitval import FitVal
+from qiskit_experiments.framework import ExperimentDecoder, FitVal
 
 ATOL_DEFAULT = 1e-2
 RTOL_DEFAULT = 1e-5
